fix: generate GraalVM native image configuration
The TypeHint annotation is used to create hints for the GraalVM native
image compiler in order to make certain classes available for
reflection. The configuration for the micronaut-graal annotation
processor was missing, which causes the GraalVM configuration to be
absent.

This change adds the micronaut-graal annotation processor to the build
configuration in order to generate the GraalVM configuration and include
it in the resulting artifacts.