Fix BOM generation
By using a version declared in a catalog, we will consistently
generate a `rxjava2` entry in the generated catalog. This is
important because when `micronaut-core` will import the catalog,
it will _inline_ entries of this catalog, so that an entry
for `mn.rxjava2` exists.

This fixes the error seen at https://github.com/micronaut-projects/micronaut-tracing/runs/6651567986?check_suite_focus=true#step:8:63